+++ src/api/config.json
@@ -1,0 +1,3 @@
+{
+    "BASE_URL": "https://localhost:7166/api/"
+}
