+++ src/api/services/StoreService.js
@@ -1,4 +1,4 @@
-import { ApiQueryGet } from "./Api";
+import { ApiQueryGet } from "../Api";
 
 const GetAllStores = async () => {
     const data = await ApiQueryGet("StoreService/GetAllStoresAsync")
